# gym-backend-api/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Weather API",
    "description": "JSON REST API for viewing weather data"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/users/image/{userID}/{imagePath}": {
      "put": {
        "summary": "Updates the user image path",
        "description": "",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the user"
          },
          {
            "name": "imagePath",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path of the image"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Updates the user image path",
          "content": {
            "application/json": {
              "schema": {
                "imagePath": "images/users/1.jpg",
                "userID": "1"
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Logs in the user",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {}
      }
    },
    "/users/logout": {
      "post": {
        "summary": "Logs out the user",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Logs out the user",
          "content": {
            "application/json": {
              "schema": {
                "authenticationKey": "88f76721-ff89-46f2-ac3a-c0c8a24d5625"
              },
              "example": {
                "authenticationKey": "88f76721-ff89-46f2-ac3a-c0c8a24d5625"
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get all users",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "description": "Gets all the users"
        }
      },
      "post": {
        "summary": "Create a user",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Creates a user",
          "content": {
            "application/json": {
              "schema": {
                "user_email": "string",
                "user_password": "string",
                "user_role": "string",
                "user_phone": "string",
                "user_first_name": "string",
                "user_last_name": "string,",
                "user_address": "string"
              },
              "example": {
                "user_email": "student@email.com",
                "user_password": "abc123",
                "user_role": "student",
                "user_phone": "1234567890",
                "user_first_name": "test",
                "user_last_name": "test",
                "user_address": "123 test street"
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Updates a user by id",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Updates a user by id",
          "content": {
            "application/json": {
              "schema": {
                "user_id": "string",
                "user_email": "string",
                "user_password": "string",
                "user_role": "string",
                "user_phone": "string",
                "user_first_name": "string",
                "user_last_name": "string,",
                "user_address": "string"
              },
              "example": {
                "user_id": "1",
                "user_email": "trainer@email.com",
                "user_password": "abc123",
                "user_role": "trainer",
                "user_phone": "1234567890",
                "user_first_name": "test",
                "user_last_name": "test",
                "user_address": "123 test street"
              }
            }
          }
        }
      }
    },
    "/users/trainers": {
      "get": {
        "summary": "Get all trainers",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "description": "Gets all the trainers"
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get a user by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Gets a user by id"
        }
      },
      "delete": {
        "summary": "Delete a user by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/by-key/{authenticationKey}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authenticationKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "summary": "Register a user",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Registers a user",
          "content": {
            "application/json": {
              "schema": {
                "user_email": "string",
                "user_password": "string",
                "user_phone": "string",
                "user_first_name": "string",
                "user_last_name": "string",
                "user_address": "string",
                "user_image_path": "string"
              },
              "example": {
                "user_email": "member@email.com",
                "user_password": "abc123",
                "user_phone": "1234567890",
                "user_first_name": "test",
                "user_last_name": "test",
                "user_address": "123 test street",
                "user_image_path": "default_user.png"
              }
            }
          }
        }
      }
    },
...
  }
}